requirements.txt
fastapi
uvicorn
passlib[bcrypt]
sqlalchemy
pydantic
python-jose
pydantic[email]
python-multipart